+++ macros/submit_merge.sh
@@ -9,31 +9,39 @@ fi
 
 source setup_env.sh
 
-runnumber=$1
+listfile=$1
 
-retDir=${PWD}
+if [ ! -f $listfile ]; then
+    echo "Nothing"
+    exit 0;
+fi
 
-cFile=mbdana24_dlis.job
+cFile=mbdana_merge_dlis.job
 DATE=`date +%Y%m%d`
 TIME=`date +%H%M%S`
-cDir=condorDir/condor_"$DATE"_"$TIME"
+cDir=condorDir/
 
+rm -rf $cDir
 mkdir -p $cDir
 
 cFile2=${cFile%.job}
 cFile2="$cFile2"_"$DATE"_"$TIME".job
+
 cp ./job_files/$cFile $cFile2
 
-sed -i -e "s@INITDIR@$PWD/$cDir@g" $cFile2
+sed -i -e "s@INITDIR@${PWD}/${cDir}@g" $cFile2
 
-echo "Queue ${segments}" >> $cFile2
+echo "Queue runnumber from ${listfile}" >> $cFile2
 
 cp $cFile2 $cDir
 rm $cFile2
-cp scriptsforcondor/scriptForMbdAna24.sh $cDir
+cp scriptsforcondor/finish_MbdAna.sh $cDir
 cp setup_env.sh $cDir
-sed -i -e "s@RUN@$runnumber@g" $cDir/scriptForMbdAna24.sh
 
+redir=${PWD}
+echo ${redir}
 cd $cDir
+pwd
 condor_submit $cFile2
-cd $retDir
+cd $redir
+pwd
